Making some progress

+++ src/core/siril.h
@@ -400,7 +400,7 @@ typedef struct {
 	WORD hi;		// MIPS-HI key in FITS file, "Upper visualization cutoff"
 	double data_max;	// used to check if 32b float is in the [0, 1] range
 	double data_min;	// used to check if 32b float is in the [0, 1] range
-	float pixel_size_x, pixel_size_y;	// XPIXSZ and YPIXSZ keys
+	double pixel_size_x, pixel_size_y;	// XPIXSZ and YPIXSZ keys
 	unsigned int binning_x, binning_y;	// XBINNING and YBINNING keys
 	char row_order[FLEN_VALUE];
 	GDateTime *date, *date_obs;		// creation and acquisition UTC dates
